requirements.txt
discord.py>=2.3.2
anthropic>=0.40.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
supabase>=2.3.0
openai>=1.40.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
aiohttp>=3.9.0
